Add KATI_DEBUG_FLAG for embedded kati running
For ckati is commonly launched from soong_ui while Android is being
built, we cannot manually run it in console. Thus, we add a new env
variable KATI_DEBUG_FLAG to control whether to start debug session.

TODO: stdin/out to tty do not work for ckati in common cases. Client
is to be created for this case.